Fix load config
